chore: update uniswapv3 oracle test
